pyproject.toml
[project]
name = "django-react"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergiu Frunză",email = "sergiu.frunza120@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "levenshtein (>=0.27.1,<0.28.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",

]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"